pyproject.toml
[tool.poetry]
name = "лабораторні роботи"
version = "0.1.0"
description = ""
authors = ["Ivan <ivanbozhkod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.3"
pillow = "^11.0.0"
matplotlib = "^3.9.3"
scipy = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
